Break client up
